C#プロジェクト内の温度情報のushort型をfloat型に変更してみたが，うまくいかなかった

+++ ThermalProfiler/ThermalProfiler/MainWindow.xaml.cs
@@ -279,7 +279,7 @@ namespace ThermalProfiler
             return temp;
         }
 
-        public static double ConvertToTemp(ushort data)
+        public static double ConvertToTemp(float data)
         {
             return (data - 1000.0) / 10.0;
         }
