+++ ThermalProfiler/libirimagerNet/ThermalPaletteImage.cs
@@ -4,13 +4,13 @@ namespace libirimagerNet
 {
     public class ThermalPaletteImage
     {
-        public ThermalPaletteImage(ushort[,] thermalImage, Bitmap paletteImage)
+        public ThermalPaletteImage(float[,] thermalImage, Bitmap paletteImage)
         {
             ThermalImage = thermalImage;
             PaletteImage = paletteImage;
         }
 
-        public ushort[,] ThermalImage { get; }
+        public float[,] ThermalImage { get; }
 
         public Bitmap PaletteImage { get; }
     }
